refactor: better magagement of bytes
